# contracts/openapi/model2.yaml
openapi: "3.1.0"
info:
  title: "Sentinel Model 2 — Unified Viewing & ANPR Analytics API"
  description: |
    Unified CCTV viewing platform connecting directly to departmental RTSP streams
    with ANPR-based metadata generation, real-time watchlist matching, and
    searchable vehicle movement records.

    **Key capabilities:**
    - Live RTSP stream catalogue from Sentinel grid (`/api/ingest` compatible)
    - ANPR (Automatic Number Plate Recognition) via PaddleOCR
    - Vehicle detection via YOLOv8n (CPU)
    - Real-time watchlist cross-referencing
    - Event tagging and searchable movement history (OpenSearch)
    - HLS relay for browser playback
  version: "1.0.0"
servers:
  - url: "http://localhost:8002"
    description: "Local development"
security:
  - BearerAuth: []
...
paths:
  /api/v1/streams:
    get:
      operationId: listStreams
      tags: [streams]
      summary: List all available camera streams
      description: |
        Returns the stream catalogue, compatible with the Sentinel `/api/ingest` format.
        Includes RTSP, WebRTC (WHEP), and HLS endpoints for each camera.
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [live, offline, connecting, error]
        - name: department_id
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Stream catalogue
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StreamCatalogueResponse" }
  /api/v1/streams/{stream_id}:
    parameters:
      - name: stream_id
        in: path
        required: true
        schema: { type: string }
    get:
      operationId: getStream
      tags: [streams]
      summary: Get stream details and connection URLs
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StreamDetail" }
          description: Stream detail
  /api/v1/streams/{stream_id}/connect:
    parameters:
      - name: stream_id
        in: path
        required: true
        schema: { type: string }
    post:
      operationId: connectStream
      tags: [streams]
      summary: Activate RTSP consumer for a stream
      description: Starts the RTSP TCP consumer with analytics pipeline
      responses:
        "200":
          description: Consumer started
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StreamConnectResponse" }
  /api/v1/streams/{stream_id}/disconnect:
    parameters:
      - name: stream_id
        in: path
        required: true
        schema: { type: string }
    post:
      operationId: disconnectStream
      tags: [streams]
      summary: Stop RTSP consumer for a stream
      responses:
        "200":
          description: Consumer stopped
...
  /api/v1/watchlist:
    get:
      operationId: listWatchlist
      tags: [watchlist]
      summary: List watchlist entries
      parameters:
        - name: type
          in: query
          schema: { $ref: "#/components/schemas/WatchlistType" }
        - name: active
          in: query
          schema: { type: boolean, default: true }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: page_size
          in: query
          schema: { type: integer, default: 50 }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WatchlistListResponse" }
          description: Watchlist

    post:
      operationId: addToWatchlist
      tags: [watchlist]
      summary: Add entry to watchlist
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WatchlistCreate" }
      responses:
        "201":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WatchlistEntry" }
          description: Created entry
  /api/v1/watchlist/{entry_id}:
    parameters:
      - name: entry_id
        in: path
        required: true
        schema: { type: string, format: uuid }
    delete:
      operationId: removeFromWatchlist
      tags: [watchlist]
      summary: Remove entry from watchlist
      responses:
        "204":
          description: Removed
  /api/v1/watchlist/alerts:
    get:
      operationId: listWatchlistAlerts
      tags: [watchlist, alerts]
      summary: List watchlist match alerts
      parameters:
        - name: acknowledged
          in: query
          schema: { type: boolean }
        - name: from_time
          in: query
          schema: { type: string, format: date-time }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: page_size
          in: query
          schema: { type: integer, default: 50 }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AlertListResponse" }
          description: Alert list
  /api/v1/watchlist/alerts/{alert_id}/acknowledge:
    parameters:
      - name: alert_id
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      operationId: acknowledgeAlert
      tags: [watchlist, alerts]
      summary: Acknowledge a watchlist alert
      responses:
        "200":
          description: Alert acknowledged

  # ── Events ────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    StreamStatus:
      type: string
      enum: [live, offline, connecting, reconnecting, error]
    WatchlistType:
      type: string
      enum: [stolen_vehicle, wanted_person, missing_person, blacklisted_vehicle, suspect]
...
    StreamDetail:
      type: object
      required: [id, camera_id, status, rtsp_url, hls_url]
      properties:
        id:
          type: string
        camera_id:
          type: string
        name:
          type: string
        status:
          $ref: "#/components/schemas/StreamStatus"
        rtsp_url:
          type: string
          format: uri
          description: "RTSP TCP stream URL (requires rtsp_transport=tcp)"
        hls_url:
          type: string
          format: uri
          description: HLS playlist URL (browser compatible)
        webrtc_url:
          type: string
          format: uri
          description: "WebRTC WHEP URL (low latency browser)"
        codec:
          type: string
        resolution:
          type: string
        frame_rate:
          type: number
          description: Declared frame rate (use PTS for timing, not this value)
        bitrate_kbps:
          type: integer
        location:
          type: object
          properties:
            latitude: { type: number }
            longitude: { type: number }
            district: { type: string }
        analytics_active:
          type: boolean
          description: Whether ANPR/detection pipeline is running for this stream
        last_frame_at:
          type: string
          format: date-time
          nullable: true
    StreamCatalogueResponse:
      type: object
      properties:
        streams:
          type: array
          items: { $ref: "#/components/schemas/StreamDetail" }
        total:
          type: integer
        active_count:
          type: integer
...
    StreamConnectResponse:
      type: object
      properties:
        stream_id:
          type: string
        status:
          $ref: "#/components/schemas/StreamStatus"
        analytics_pipeline:
          type: boolean
        message:
          type: string
...
    WatchlistCreate:
      type: object
      required: [type, identifier]
      properties:
        type:
          $ref: "#/components/schemas/WatchlistType"
        identifier:
          type: string
          description: Plate number, person name, or other identifier
        description:
          type: string
        case_number:
          type: string
          description: Police case/FIR number
        priority:
          type: string
          enum: [low, medium, high, critical]
          default: medium
        source_db:
          type: string
          enum: [vahan, sarthi, egujcop, afis, nafis, manual]
          default: manual
        source_id:
          type: string
          description: ID in the source database
        metadata:
          type: object
          additionalProperties: true
        expires_at:
          type: string
          format: date-time
          nullable: true
    WatchlistEntry:
      allOf:
        - $ref: "#/components/schemas/WatchlistCreate"
        - type: object
          properties:
            id:
              type: string
              format: uuid
            active:
              type: boolean
            hit_count:
              type: integer
            last_hit_at:
              type: string
              format: date-time
              nullable: true
            created_at:
              type: string
              format: date-time
            created_by:
              type: string
    WatchlistListResponse:
      type: object
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/WatchlistEntry" }
        total:
          type: integer
    WatchlistAlert:
      type: object
      properties:
        id:
          type: string
          format: uuid
        watchlist_entry_id:
          type: string
          format: uuid
        detection_id:
          type: string
          format: uuid
        camera_id:
          type: string
        plate_number:
          type: string
        alert_type:
          $ref: "#/components/schemas/WatchlistType"
        priority:
          type: string
          enum: [low, medium, high, critical]
        triggered_at:
          type: string
          format: date-time
        acknowledged:
          type: boolean
        acknowledged_by:
          type: string
          nullable: true
        acknowledged_at:
          type: string
          format: date-time
          nullable: true
        location:
          type: object
          properties:
            latitude: { type: number }
            longitude: { type: number }
            district: { type: string }
        snapshot_url:
          type: string
          format: uri
          nullable: true
    AlertListResponse:
      type: object
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/WatchlistAlert" }
        total:
          type: integer
        unacknowledged_count:
          type: integer
...